:lipstick: style: make algorithms better
